Move loadFile from file-utils.ts to file-handlers.ts
